+++ src/actors/constants.ts
@@ -1,0 +1,11 @@
+import { Color } from "excalibur";
+
+export const ShipColors = [
+  Color.DarkGray,
+  Color.LightGray,
+  Color.Azure,
+  Color.Magenta
+];
+
+export const radius = 2;
+export const ShipNames = ["Trader", "Pirate"];
